feat: serialize signature
